Renamed ILogger -> Logger
