+++ cdoc/Recipient.cpp
@@ -21,7 +21,6 @@
 #include "CDoc2.h"
 #include "Certificate.h"
 #include "Crypto.h"
-#include "ILogger.h"
 #include "Utils.h"
 
 #include <algorithm>
